Move test-ui.ts to /query-testing/test-ui.ts
